% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRASSVis-package.R
\docType{package}
\name{BRASSVis-package}
\alias{BRASSVis}
\alias{BRASSVis-package}
\title{BRASSVis: Visualization of Structural Variants from BRASS}
\description{
Visualize gene fusions and structural variants detected by BRASS (BReakpoint AnalySiS) from whole-genome sequencing data. Reads BEDPE files and produces publication-ready PDF reports showing breakpoints, exon structure, ideograms, circos plots, and protein domain retention. Based on the Arriba visualization approach adapted for WGS structural variant data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bczech/BRASSVis}
  \item Report bugs at \url{https://github.com/bczech/BRASSVis/issues}
}

}
\author{
\strong{Maintainer}: Bartosz Czech \email{bartosz.w.czech@gmail.com} (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})

Authors:
\itemize{
  \item Bartosz Czech \email{bartosz.w.czech@gmail.com} (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
}

}
\keyword{internal}
